pyproject.toml
[tool.poetry]
name = "balloon-sim"
version = "0.1.0"
description = "Simulate the path traveled by a balloon using real data"
authors = ["Cliff Yuan Li <cliff.li@netlight.com>"]
packages = [
    { include = "weather_api", from = "src" },
    { include = "simulation", from = "src" },
    { include = "animation", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
matplotlib = "^3.7.1"
cartopy = "^0.21.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.1.1"
pylint = "^2.17.2"
pytest = "^7.2.2"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
